@@ -2,7 +2,7 @@
   "name": "pepeverde/libs",
   "description": "Librerie per lo sviluppo di applicazioni",
   "type": "library",
-  "version": "0.18.0",
+  "version": "0.19.0",
   "license": "MIT",
   "authors": [
     {
@@ -15,27 +15,27 @@
     }
   ],
   "require": {
-    "php": "7.2.*|7.3.*|7.4.*",
+    "php": "^8.1",
     "ext-fileinfo": "*",
     "ext-imagick": "*",
     "ext-mbstring": "*",
     "lib-pcre": "*",
-    "aura/session": "~2.1",
-    "behat/transliterator": "~1.3",
-    "endelwar/zigra": "0.9.*",
+    "aura/session": "~4.0",
+    "behat/transliterator": "^1.5.0",
+    "endelwar/zigra": "0.10.*",
     "imagecow/imagecow": "^2.4",
-    "pelago/emogrifier": "~v4.0.0",
-    "pepeverde/doctrine1": "^1.3.16",
+    "pelago/emogrifier": "^7.0",
+    "diablomedia/doctrine1": "^4.0.4",
     "sentry/sdk": "3.1.*",
-    "soundasleep/html2text": "~1.1",
+    "soundasleep/html2text": "~2.1",
     "swiftmailer/swiftmailer": "v6.3.0",
-    "twig/twig": "^2.14|^3",
-    "vlucas/phpdotenv": "~2.6"
+    "twig/twig": "^v3.17.1",
+    "vlucas/phpdotenv": "^5.6.1"
   },
   "require-dev": {
-    "friendsofphp/php-cs-fixer": "^3.0",
-    "phpstan/phpstan": "^1.2",
-    "phpunit/phpunit": "^8.5"
+    "friendsofphp/php-cs-fixer": "^3.65",
+    "phpstan/phpstan": "^1.12",
+    "phpunit/phpunit": "^9.6"
   },
   "autoload": {
     "psr-4": {
@@ -49,6 +49,9 @@
   },
   "config": {
     "sort-packages": true,
-    "optimize-autoloader": true
+    "optimize-autoloader": true,
+    "allow-plugins": {
+      "php-http/discovery": true
+    }
   }
 }
