@@ -2,8 +2,6 @@
 
 namespace Pepeverde;
 
-use Zigra_Exception;
-
 class Error
 {
     /**
@@ -15,7 +13,7 @@ class Error
         string $appVersion,
         string $environment,
         array $extra,
-        bool $send_default_pii
+        bool $send_default_pii,
     ): void {
         \Sentry\init([
             'dsn' => $sentryConfig['sentry-server'],
@@ -24,7 +22,7 @@ class Error
             'send_default_pii' => $send_default_pii,
         ]);
 
-        \Sentry\configureScope(static function (\Sentry\State\Scope $scope) use ($extra): void {
+        \Sentry\configureScope(static function(\Sentry\State\Scope $scope) use ($extra): void {
             $scope->setExtras($extra);
         });
     }
@@ -34,13 +32,13 @@ class Error
      * @param array<string, mixed>      $extra
      */
     public static function enableErrorHandler(
-        array $sentryConfig = null,
+        ?array $sentryConfig = null,
         string $appVersion = 'dev',
         string $environment = 'development',
         array $extra = [],
-        bool $send_default_pii = false
+        bool $send_default_pii = false,
     ): void {
-        if (null === $sentryConfig || (is_countable($sentryConfig) && 0 === count($sentryConfig))) {
+        if (null === $sentryConfig || 0 === count($sentryConfig)) {
             throw new \RuntimeException('No Sentry configuration available');
         }
 
@@ -53,7 +51,7 @@ class Error
     public static function report(\Exception $e, bool $display = true, array $data = []): void
     {
         if (!empty($data)) {
-            \Sentry\configureScope(static function (\Sentry\State\Scope $scope) use ($data): void {
+            \Sentry\configureScope(static function(\Sentry\State\Scope $scope) use ($data): void {
                 $scope->setExtras($data);
             });
         }
@@ -65,7 +63,7 @@ class Error
             $message .= ' [Event-ID: ' . $eventId . ']';
         }
         if (true === $display) {
-            Zigra_Exception::renderError(
+            \Zigra_Exception::renderError(
                 $e->getCode(),
                 $message
             );
@@ -77,9 +75,9 @@ class Error
         \Sentry\captureMessage($message);
     }
 
-    public static function setUser(string $username, string $email = null): void
+    public static function setUser(string $username, ?string $email = null): void
     {
-        \Sentry\configureScope(static function (\Sentry\State\Scope $scope) use ($username, $email): void {
+        \Sentry\configureScope(static function(\Sentry\State\Scope $scope) use ($username, $email): void {
             $scope->setUser(
                 \Sentry\UserDataBag::createFromArray(
                     [
