@@ -2,14 +2,11 @@
 
 namespace Pepeverde;
 
-use finfo;
-use SplFileInfo;
-
-class FileInfo extends SplFileInfo
+class FileInfo extends \SplFileInfo
 {
     public function getMimeType(): string
     {
-        $finfo = new finfo();
+        $finfo = new \finfo();
 
         return $finfo->file($this->getRealPath());
     }
