@@ -2,32 +2,19 @@
 
 namespace Pepeverde;
 
-use Exception;
-use FilesystemIterator;
 use Imagecow\Image;
 use Imagecow\ImageException;
-use ImagickException;
-use RecursiveDirectoryIterator;
-use RecursiveIteratorIterator;
-use RuntimeException;
 
 class FileSystem
 {
-    /**
-     * @param string $dir
-     */
-    public static function isDirEmpty($dir): bool
+    public static function isDirEmpty(string $dir): bool
     {
-        $di = new RecursiveDirectoryIterator($dir, FilesystemIterator::SKIP_DOTS);
+        $di = new \RecursiveDirectoryIterator($dir, \FilesystemIterator::SKIP_DOTS);
 
         return 0 === iterator_count($di);
     }
 
-    /**
-     * @param string $name
-     * @param int    $retinaType
-     */
-    public static function getRetinaName($name, $retinaType = 2): ?string
+    public static function getRetinaName(string $name, int $retinaType = 2): ?string
     {
         if (!empty($name)) {
             $v = pathinfo($name);
@@ -38,12 +25,7 @@ class FileSystem
         return null;
     }
 
-    /**
-     * @param string $path
-     *
-     * @return bool|string
-     */
-    public static function getFileTypeFromPath($path)
+    public static function getFileTypeFromPath(string $path): bool|string
     {
         if (!is_file($path)) {
             return false;
@@ -52,46 +34,26 @@ class FileSystem
         return self::getFileTypeFromMime(mime_content_type($path));
     }
 
-    /**
-     * @param string $mime
-     */
-    private static function getFileTypeFromMime($mime): string
+    private static function getFileTypeFromMime(string $mime): string
     {
-        switch ($mime) {
-            case 'application/pdf':
-                return 'pdf';
-            case 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet':
-            case 'application/vnd.openxmlformats-officedocument.spreadsheetml.template':
-            case 'application/vnd.ms-excel':
-                return 'excel';
-            case 'application/vnd.openxmlformats-officedocument.wordprocessingml.document':
-            case 'application/vnd.openxmlformats-officedocument.wordprocessingml.template':
-            case 'application/msword':
-                return 'word';
-            case 'application/zip':
-            case 'application/x-rar-compressed':
-            case 'application/x-7z-compressed':
-                return 'archive';
-            case 'image/jpeg':
-            case 'image/png':
-                return 'image';
-            default:
-                return 'file';
-        }
+        return match ($mime) {
+            'application/pdf' => 'pdf',
+            'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet', 'application/vnd.openxmlformats-officedocument.spreadsheetml.template', 'application/vnd.ms-excel' => 'excel',
+            'application/vnd.openxmlformats-officedocument.wordprocessingml.document', 'application/vnd.openxmlformats-officedocument.wordprocessingml.template', 'application/msword' => 'word',
+            'application/zip', 'application/x-rar-compressed', 'application/x-7z-compressed' => 'archive',
+            'image/jpeg', 'image/png' => 'image',
+            default => 'file',
+        };
     }
 
-    /**
-     * @param int|float|string|null $size
-     * @param int                   $precision
-     */
-    public static function formatSize($size, $precision = 2): string
+    public static function formatSize(float|int|string|null $size, int $precision = 2): string
     {
         if (!is_numeric($size)) {
-            return '0';
+            return '0.00';
         }
 
         if (0 === $size || '0' === $size) {
-            return '0';
+            return '0.00';
         }
 
         $size = (float)$size;
@@ -99,15 +61,15 @@ class FileSystem
         $base = log($size) / log(1024);
         $suffixes = ['B', 'kB', 'MB', 'GB', 'TB', 'PB'];
 
-        return round(1024 ** ($base - floor($base)), $precision) . $suffixes[(int)floor($base)];
+        return number_format(1024 ** ($base - floor($base)), $precision) . $suffixes[(int)floor($base)];
     }
 
-    public static function deleteDir($path): void
+    public static function deleteDir(string $path): void
     {
         if (is_dir($path)) {
-            $iterator = new RecursiveIteratorIterator(
-                new RecursiveDirectoryIterator($path, RecursiveDirectoryIterator::SKIP_DOTS),
-                RecursiveIteratorIterator::CHILD_FIRST
+            $iterator = new \RecursiveIteratorIterator(
+                new \RecursiveDirectoryIterator($path, \RecursiveDirectoryIterator::SKIP_DOTS),
+                \RecursiveIteratorIterator::CHILD_FIRST
             );
 
             foreach ($iterator as $file) {
@@ -124,35 +86,40 @@ class FileSystem
         }
     }
 
-    public static function deleteFile($path_to_file): void
+    public static function deleteFile(string $path_to_file): void
     {
         if (is_file($path_to_file)) {
             unlink($path_to_file);
         }
     }
 
+    /**
+     * @param array{width: int, height: int} $uploaded_dimensions
+     * @param array{width: int, height: int} $required_image_dimensions
+     */
     public static function checkImageDimensions(array $uploaded_dimensions, array $required_image_dimensions): bool
     {
         if (!isset($uploaded_dimensions['width'], $uploaded_dimensions['height'], $required_image_dimensions['width'], $required_image_dimensions['height'])) {
-            throw new RuntimeException('Dimensions must be set to be checked');
+            throw new \RuntimeException('Dimensions must be set to be checked');
         }
 
         return ($uploaded_dimensions['width'] >= $required_image_dimensions['width']) && ($uploaded_dimensions['height'] >= $required_image_dimensions['height']);
     }
 
+    /**
+     * @param array{width: int, height: int} $uploaded_dimensions
+     * @param array{width: int, height: int} $required_image_dimensions
+     */
     public static function checkFixedImageDimensions(array $uploaded_dimensions, array $required_image_dimensions): bool
     {
         if (!isset($uploaded_dimensions['width'], $uploaded_dimensions['height'], $required_image_dimensions['width'], $required_image_dimensions['height'])) {
-            throw new RuntimeException('Dimensions must be set to be checked');
+            throw new \RuntimeException('Dimensions must be set to be checked');
         }
 
         return ((int)$uploaded_dimensions['width'] === (int)$required_image_dimensions['width']) && ((int)$uploaded_dimensions['height'] === (int)$required_image_dimensions['height']);
     }
 
-    /**
-     * @param string $image_type
-     */
-    public static function checkIfIsImage($image_type, array $allowed_types): bool
+    public static function checkIfIsImage(string $image_type, array $allowed_types): bool
     {
         if (in_array($image_type, $allowed_types, true)) {
             return true;
@@ -162,10 +129,11 @@ class FileSystem
     }
 
     /**
-     * @param Image $image
-     * @param array $uploaded_dimensions
+     * @param array{width: int, height: int} $uploaded_dimensions
+     *
+     * @return array{width: int, height: int}
      */
-    public static function setDimensionsVariables(&$image, $uploaded_dimensions): array
+    public static function setDimensionsVariables(Image $image, array $uploaded_dimensions): array
     {
         $uploaded_check = [
             'width' => 0,
@@ -183,17 +151,26 @@ class FileSystem
     }
 
     /**
-     * @param string $path_to_upload
+     * @param array{width: int, height: int} $required_dimensions
+     * @param array{width: int, height: int} $uploaded_dimensions
      */
-    public static function uploadAndVerifyImage(array $post_files, $path_to_upload, array &$error, array $required_dimensions, array $uploaded_dimensions, array $allowed_types): array
-    {
+    public static function uploadAndVerifyImage(
+        array $post_files,
+        string $path_to_upload,
+        array &$error,
+        array $required_dimensions,
+        array $uploaded_dimensions,
+        array $allowed_types,
+        ?FileMover $fileMover = null,
+    ): array {
+        $fileMover = $fileMover ?? new FileMover();
         $image_name = null;
-        if (4 !== $post_files['error']) {
+        if (\UPLOAD_ERR_NO_FILE !== $post_files['error']) {
             try {
-                if ($image_name = Upload::uploadFile($post_files, $path_to_upload, $post_files['name'])) {
+                if ($image_name = Upload::uploadFile($post_files, $path_to_upload, $post_files['name'], $fileMover)) {
                     $image = Image::fromFile($path_to_upload . '/' . $image_name);
                     if (self::checkIfIsImage($image->getMimeType(), $allowed_types)) {
-                        //setto le dimensioni richieste per l'immagine
+                        // setto le dimensioni richieste per l'immagine
                         $uploaded_check = self::setDimensionsVariables($image, $uploaded_dimensions);
                         if (self::checkImageDimensions($uploaded_check, $required_dimensions)) {
                             return [
@@ -205,15 +182,19 @@ class FileSystem
                     } else {
                         $error[] = 'il file caricato non è un\'immagine, file non caricato.';
                     }
+                } else {
+                    $error[] = 'Upload::uploadFile returned false';
                 }
             } catch (ImageException $e) {
                 $error[] = 'il file caricato non è un\'immagine valida, file non caricato.';
-            } catch (ImagickException $e) {
+            } catch (\ImagickException $e) {
                 $error[] = 'il file caricato non è un\'immagine valida, file non caricato.';
-            } catch (Exception $e) {
+            } catch (\Exception $e) {
                 $error[] = $e->getMessage();
                 Error::report($e, false);
             }
+        } else {
+            $error[] = 'No file was uploaded';
         }
 
         return [
@@ -224,17 +205,26 @@ class FileSystem
     }
 
     /**
-     * @param string $path_to_upload
+     * @param array{width: int, height: int} $required_dimensions
+     * @param array{width: int, height: int} $uploaded_dimensions
      */
-    public static function uploadAndVerifyFixedImage(array $post_files, $path_to_upload, array &$error, array $required_dimensions, array $uploaded_dimensions, array $allowed_types): array
-    {
+    public static function uploadAndVerifyFixedImage(
+        array $post_files,
+        string $path_to_upload,
+        array &$error,
+        array $required_dimensions,
+        array $uploaded_dimensions,
+        array $allowed_types,
+        ?FileMover $fileMover = null,
+    ): array {
+        $fileMover = $fileMover ?? new FileMover();
         $image_name = null;
         if (4 !== $post_files['error']) {
             try {
                 if ($image_name = Upload::uploadFile($post_files, $path_to_upload, $post_files['name'])) {
                     $image = Image::fromFile($path_to_upload . '/' . $image_name);
                     if (self::checkIfIsImage($image->getMimeType(), $allowed_types)) {
-                        //setto le dimensioni richieste per l'immagine
+                        // setto le dimensioni richieste per l'immagine
                         $uploaded_check = self::setDimensionsVariables($image, $uploaded_dimensions);
                         if (self::checkFixedImageDimensions($uploaded_check, $required_dimensions)) {
                             return [
@@ -249,9 +239,9 @@ class FileSystem
                 }
             } catch (ImageException $e) {
                 $error[] = 'il file caricato non è un\'immagine valida, file non caricato.';
-            } catch (ImagickException $e) {
+            } catch (\ImagickException $e) {
                 $error[] = 'il file caricato non è un\'immagine valida, file non caricato.';
-            } catch (Exception $e) {
+            } catch (\Exception $e) {
                 $error[] = $e->getMessage();
                 Error::report($e, false);
             }
@@ -265,15 +255,16 @@ class FileSystem
     }
 
     /**
-     * @param Image  $image
-     * @param array  $cropdata
-     * @param string $savepath
-     * @param string $image_name
+     * @param array<string, int> $cropdata
      */
-    public static function cropImage($image, $cropdata, $savepath, $image_name): Image
-    {
+    public static function cropImage(
+        Image $image,
+        array $cropdata,
+        string $savepath,
+        string $image_name,
+    ): Image {
         if (!file_exists($savepath) && !mkdir($savepath, 0777, true) && !is_dir($savepath)) {
-            throw new RuntimeException(sprintf('Directory "%s" was not created', $savepath));
+            throw new \RuntimeException(sprintf('Directory "%s" was not created', $savepath));
         }
 
         $image->crop($cropdata['width'], $cropdata['height'], $cropdata['x'], $cropdata['y'])
