@@ -5,12 +5,9 @@ namespace Pepeverde;
 class Helper
 {
     /**
-     * @param ?string $key
-     * @param ?mixed  $default
-     *
      * @return array|bool|mixed|string|null
      */
-    public static function env($key, $default = null)
+    public static function env(?string $key, mixed $default = null)
     {
         $value = getenv($key);
         if (false === $value) {
