@@ -8,10 +8,10 @@ use PHPUnit\Framework\TestCase;
 
 class HelperTest extends TestCase
 {
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
-        $dotenv = new Dotenv(__DIR__ . '/../resources/env/');
+        $dotenv = Dotenv::createUnsafeImmutable(__DIR__ . '/../resources/env/');
         $dotenv->load();
     }
 
